pyproject.toml
[project]
name = "kinitro-agent"
version = "0.0.1"
description = "Kinitro Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "metaworld>=3.0.0",
    "torch>=2.8.0",
    "gymnasium>=0.29.0",
    "mujoco>=3.0.0",
    "numpy>=1.24.0",
    "pycapnp>=2.1.0",
    "tensorboard>=2.15.0",
    "matplotlib>=3.7.0",
    "distrax>=0.1.5",
    "flax>=0.9.0",
    "jaxtyping>=0.2.34",
    "numpy>=2.1.2",
    "orbax-checkpoint<=0.11.11",
    "tyro>=0.8.11",
    "scipy>=1.15.2",
]

[dependency-groups]
dev = ["debugpy>=1.8.9", "py-spy>=0.4.0", "pytest>=8.3.4", "ruff>=0.8.2"]

[project.optional-dependencies]
cpu = ["jax<0.7.0"]
# https://github.com/jax-ml/jax/issues/27062
metal = ["jax<=0.5.0", "jax-metal>=0.1.0; sys_platform == 'darwin'"]
cuda12 = ["jax[cuda12]<0.7.0"]
tpu = ["jax[tpu]<0.7.0"]
testing = [
    "chex>=0.1.89",
    "pytest>=8.3.5",
]

[tool.uv]
prerelease = "allow"
conflicts = [
  [
    { extra = "cpu" },
    { extra = "metal" },
    { extra = "cuda12" },
    { extra = "tpu" },
  ],
]

[tool.ruff]
ignore = ["F722"]
ignore-init-module-imports = true
